+++ test/unit/compiler/concerns/compiles_statements_test.rb
@@ -28,21 +28,34 @@ class CompilesStatementsTest < TestCase
     assert_compiles_to "@not_a_real_directive   (1, 2, 3)", "_out<<'@not_a_real_directive   (1, 2, 3)';"
   end
 
-  class CustomDirectiveHandler
-    def custom_directive args
-      return args&.join(",") || 'no arguments'
+  def test_register_directive_handler
+    RBlade::register_directive_handler('custom_directive') { 'simple' }
+    RBlade::register_directive_handler('custom_directive_with_args') { |args| args&.join(",") || 'no arguments' }
+    RBlade::register_directive_handler('custom_directive_with_tokens') do |tokens, token_index|
+      tokens[token_index].value[:name]
     end
+
+    assert_compiles_to "@custom_directive", nil, "simple"
+    assert_compiles_to "@CustomDirective", nil, "simple"
+    assert_compiles_to "@CUSTOMDIRECTIVE", nil, "simple"
+    assert_compiles_to "@customdirective()", nil, "simple"
+
+    assert_compiles_to "@custom_directive_with_args", nil, "no arguments"
+    assert_compiles_to "@custom_directive_with_args(one argument)", nil, "one argument"
+    assert_compiles_to "@custom_directive_with_args(a,b)", nil, "a,b"
+    assert_compiles_to "@custom_directive_with_args(1,2,3)", nil, "1,2,3"
+    assert_compiles_to "@custom_directive_with_args('\\'\"')", nil, "'\\'\"'"
+
+    assert_compiles_to "@custom_directive_with_tokens", nil, "custom_directive_with_tokens"
   end
 
-  def test_register_directive_handler
-    RBlade::register_directive_handler('custom_directive', CustomDirectiveHandler, :custom_directive)
+  def test_sum_directive_used_in_readme
+    RBlade::register_directive_handler('sum') do |args|
+      args.inject(0) { |sum, num| sum + num.to_i }
+    end
 
-    assert_compiles_to "@custom_directive", "no arguments"
-    assert_compiles_to "@customDirective", "no arguments"
-    assert_compiles_to "@customdirective()", "no arguments"
-
-    assert_compiles_to "@customdirective(one argument)", "one argument"
-    assert_compiles_to "@customdirective(a,b)", "a,b"
-    assert_compiles_to "@customdirective(1,2,3)", "1,2,3"
+    assert_compiles_to "@sum(1)", nil, '1'
+    assert_compiles_to "@sum(1, 2)", nil, '3'
+    assert_compiles_to "@sum (1, 2, 3)", nil, '6'
   end
 end